hotfix: ping regions twice from client
For some reason the first client-to-region requests takes an abnormally long amount of time. Might have something to do with DNS resolution or the like. Making two requests is a quick and dirty way to solve this for now. These requests cost next to nothing so it isn't a big deal. Of course, it would be better to keep track of initial requests and subsequently not ping twice.